factor into component
